refactor: cdr-> vite
