--- src/App.js
@@ -1,21 +1,0 @@
-import './App.css';
-
-
-
-
-
-
-function App() {
-  return (
-    <div>
-      {/* <PageHeader />
-      <Routes>
-        <Route path='/auth/register' element={<RegisterFrom />} />
-        <Route path='/auth/login' element={<LoginForm />} />
-        <Route path='/main' element = {<Main /> }/>
-      </Routes> */}
-    </div>
-  );
-}
-
-export default App;
