+++ src/main.jsx
@@ -1,0 +1,16 @@
+import React from 'react'
+import ReactDOM from 'react-dom/client'
+import App from './App'
+import { BrowserRouter } from 'react-router-dom'
+import { AuthProvider } from './components/context/AuthContext'
+// import './index.css';
+
+ReactDOM.createRoot(document.getElementById('root')).render(
+    <React.StrictMode>
+        <BrowserRouter>
+            {/* <AuthProvider> */}
+                <App />
+            {/* </AuthProvider> */}
+        </BrowserRouter>
+    </React.StrictMode>
+)
